Added 'robust' re-fit of coefficients using RIRLS
Not sure if this will work in practice, but maybe it will fix some of the
issues caused by missed clouds/shadows. If we miss cloud/shadow, and we
don't false positive call it a change, then our regression result will be
wrong (too high, usually) and our RMSE will be too large.

This might kill fits for normally noisy environments (ag, wetlands), but
maybe it works better for forests